Use Double for trolley - 15to17 precision

+++ WooliesX.Api/Controllers/TrolleyTotalController.cs
@@ -27,7 +27,7 @@ namespace WooliesX.Api.Controllers
         }
 
         [HttpPost]
-        public async Task<float> Post(Trolley trolley)
+        public async Task<double> Post(Trolley trolley)
         {
             return await _client.ApiResourceTrolleyCalculatorPostAsync(_config.GetValue<Guid>("ApiToken"), trolley);
         }
